Tests: Moved integration tests to different directory.
