+++ tests/data/integration/list.js
@@ -14,19 +14,19 @@ import Italic from '@ckeditor/ckeditor5-basic-styles/src/italic';
 import Underline from '@ckeditor/ckeditor5-basic-styles/src/underline';
 import Link from '@ckeditor/ckeditor5-link/src/link';
 import List from '@ckeditor/ckeditor5-list/src/list';
-import PasteFromWord from '../../src/pastefromword';
+import PasteFromWord from '../../../src/pastefromword';
 
 import { setData } from '@ckeditor/ckeditor5-engine/src/dev-utils/model';
-import { expectPaste } from '../_utils/utils';
+import { expectPaste } from '../../_utils/utils';
 
-import simple from '../_data/list/simple/input.word2016.html';
-import styled from '../_data/list/styled/input.word2016.html';
-import multiple from '../_data/list/multiple/input.word2016.html';
-import multipleCombined from '../_data/list/multiple-combined/input.word2016.html';
-import manyOneItem from '../_data/list/many-one-item/input.word2016.html';
-import heading1 from '../_data/list/heading1/input.word2016.html';
-import heading3Styled from '../_data/list/heading3-styled/input.word2016.html';
-import heading7 from '../_data/list/heading7/input.word2016.html';
+import simple from '../../_data/list/simple/input.word2016.html';
+import styled from '../../_data/list/styled/input.word2016.html';
+import multiple from '../../_data/list/multiple/input.word2016.html';
+import multipleCombined from '../../_data/list/multiple-combined/input.word2016.html';
+import manyOneItem from '../../_data/list/many-one-item/input.word2016.html';
+import heading1 from '../../_data/list/heading1/input.word2016.html';
+import heading3Styled from '../../_data/list/heading3-styled/input.word2016.html';
+import heading7 from '../../_data/list/heading7/input.word2016.html';
 
 describe( 'List – integration', () => {
 	let element, editor;
